informergen,listergen: support a single-cluster-scoped informer factory
In some cases, users will want to run against a specific cluster in kcp,
so we need to allow them to create a single-cluster-scoped factory. We
can't extract the cluster they are in from the client we get passed
without major hackery, so we use the standard shared informer, keyfunc
and indices, simply leaving out the cluster entirely.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>