re-think packages and directories in cluster_codegen.sh
kcp uses an entirely different concept of placing the packages. Whereas the examples in this repo
follow a

/ <singlecluster>          <-- all single-cluster code
/ <base outputdir>
+-- clientset/versioned/   <-- cluster-aware clientset
+-- informers/
+-- listers/

structure, but in kcp it's

/ <singlecluster>
+-- clientset/
|   +-- versioned/     <-- single-cluster clientset
|        +-- cluster/  <-- cluster-ware clientset
+-- listers
+-- informers

Placing the cluster clientset right into the singlecluster clientset, but keeping the listers and
informers standalone (since there are no single-cluster listers and informers in the kcp SDK).

Making both structures possible would either require a bunch of magic, or some explicit, IMHO easier
to understand CLI flags to control exactly what you want to have happen. I have sacrified to a degree
the concept of "naming" a clientset and the "versioned" subdirectory, also because in kcp the
cluster-aware clientset doesn't itself contain a clientset/versioned/ subdirectory.

On-behalf-of: @SAP christoph.mewes@sap.com