fix boilerplate
Since most of the new code is copied straight from Kubernetes, I felt uncomfortable to relicense it
as Copyright KCP, so instead I chose to apply a dual-copyright. Ideally the KCP one should have been
after the Kubernetes line, but that is not possible with verify_boilerplate.py.

Since the examples are kcp-only, they got their own boilerplates. It felt weird to have the
generated code be dual-copyrighted.

On-behalf-of: @SAP christoph.mewes@sap.com